backend/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Marshmallow
flask-marshmallow
marshmallow-sqlalchemy
Flask-JWT-Extended
flask-cors
python-dotenv
Flask-Limiter
